requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
gTTS
aiofiles
python-multipart
